requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
websockets>=10.0
langchain>=0.1.0
langgraph>=0.0.10
edge-tts>=6.1.9
sounddevice>=0.4.6
numpy>=1.21.0
pydantic>=2.0.0
sqlalchemy>=1.4.0
aiohttp>=3.8.0
python-multipart>=0.0.5
deepseek-api>=0.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 